Consolidate tensor_ops as a single import
