+++ shumai/tensor/index.ts
@@ -1,5 +1,4 @@
 export * from './tensor'
 export * from './tensor_ops'
-export * from './tensor_ops_gen'
 import './register_gradients'
 export type { TensorOpsInterface } from './tensor_ops_interface_gen'
