+++ shumai/tensor/tensor.ts
@@ -2,7 +2,7 @@ import { ptr, toArrayBuffer } from 'bun:ffi'
 import { fl } from '../ffi/ffi_flashlight'
 import { arrayArg } from '../ffi/ffi_bind_utils'
 import { TensorOpsInterface } from './tensor_ops_interface_gen'
-import { full } from './tensor_ops_gen'
+import { full } from './tensor_ops'
 import { gen_tensor_op_shim } from './tensor_ops_shim_gen'
 import * as ops from './tensor_ops'
 import { getStack, collectStats } from './stats'
