+++ shumai/tensor/tensor_ops.ts
@@ -1,5 +1,7 @@
 import type { Tensor } from './tensor'
 
+export * from './tensor_ops_gen'
+
 export function softmax(tensor: Tensor, axis: number): Tensor {
   const exp = tensor.exp()
   return exp.div(exp.sum([axis], true))
